Fixed some more .0s

+++ src/main/java/edu/ksu/canvas/attendance/controller/SummaryController.java
@@ -158,10 +158,10 @@ public class SummaryController extends AttendanceBaseController {
     private void addAssignmentSummaryToPage(ModelAndView page, int totalPresentDays, int totalTardyDays, int totalAbsentDays, int totalExcusedDays, int totalDays, AttendanceAssignment assignment) {
 
         if (assignment != null && totalDays != 0 && !StringUtils.isBlank(assignment.getPresentPoints())) {
-            Long presentWeight = Long.valueOf(assignment.getPresentPoints());
-            Long tardyWeight = Long.valueOf(assignment.getTardyPoints());
-            Long absentWeight = Long.valueOf(assignment.getAbsentPoints());
-            Long excusedWeight = Long.valueOf(assignment.getExcusedPoints());
+            double presentWeight = Long.valueOf(assignment.getPresentPoints());
+            double tardyWeight = Long.valueOf(assignment.getTardyPoints());
+            double absentWeight = Long.valueOf(assignment.getAbsentPoints());
+            double excusedWeight = Long.valueOf(assignment.getExcusedPoints());
             Long assignmentPoints = Long.valueOf(assignment.getAssignmentPoints());
 
             double presentMultiplier = presentWeight / 100;
